requirements.txt
langchain>=0.3
langchain-core>=0.3
langchain-openai>=0.2
pydantic>=2.0
python-dotenv>=1.0
pytest>=8.0
